extracted all the business logic and pushed it up to the container component
